use 32bit int

+++ eval/benchmark.cpp
@@ -33,7 +33,7 @@ limitations under the License.
 #include "string_util.hpp"
 #include "history.hpp"
 
-using integer = unsigned long;
+using integer = unsigned int;
 
 template<typename text, typename integer>
 int benchmark(const std::string& corpus_path)
